requirements.txt
Babel==2.4.0
certifi==2018.1.18
graphviz==0.9
h5py==2.8.0
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.2.3
Markdown==3.0.1
matplotlib==2.2.3
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.15.1
opencv-contrib-python==3.4.3.18
pandas==0.23.4
Pillow==5.1.0
protobuf==3.6.1
pydot==1.2.4
Pygments==2.2.0
pylint==1.8.3
pytz==2018.5
scipy==1.1.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
Wave==0.0.2